+++ imbrium/architectures/models.py
@@ -1,12 +1,23 @@
 import tensorflow as tf
 from tensorflow import keras
-from tensorflow.keras.layers import LSTM, Dense, Flatten, Conv1D, MaxPooling1D, Dropout, Bidirectional, TimeDistributed, GRU, SimpleRNN, RepeatVector
+from tensorflow.keras.layers import (LSTM,
+                                     Dense,
+                                     Flatten,
+                                     Conv1D,
+                                     MaxPooling1D,
+                                     Dropout,
+                                     Bidirectional,
+                                     TimeDistributed,
+                                     GRU,
+                                     SimpleRNN,
+                                     RepeatVector)
 
 
 def mlp(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates MLP model by defining all layers with activation functions,
@@ -15,15 +26,26 @@ def mlp(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
             model (object): Returns compiled Keras model.
     '''
     model = keras.Sequential()
-    model.add(Dense(50, activation='relu', input_dim=input_shape))
-    model.add(Dense(25, activation='relu'))
-    model.add(Dense(25, activation='relu'))
+    model.add(
+        Dense(
+            layer_config['layer0'][0],
+            activation=layer_config['layer0'][1],
+            input_dim=input_shape))
+    model.add(
+        Dense(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1]))
+    model.add(
+        Dense(
+            layer_config['layer2'][0],
+            activation=layer_config['layer2'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -34,6 +56,7 @@ def rnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates RNN model by defining all layers with activation functions,
@@ -42,6 +65,7 @@ def rnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -50,12 +74,19 @@ def rnn(
     model = keras.Sequential()
     model.add(
         SimpleRNN(
-            40,
-            activation='relu',
+            layer_config['layer0'][0],
+            activation=layer_config['layer0'][1],
             return_sequences=True,
             input_shape=input_shape))
-    model.add(SimpleRNN(50, activation='relu', return_sequences=True))
-    model.add(SimpleRNN(50, activation='relu'))
+    model.add(
+        SimpleRNN(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1],
+            return_sequences=True))
+    model.add(
+        SimpleRNN(
+            layer_config['layer2'][0],
+            activation=layer_config['layer2'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -66,6 +97,7 @@ def lstm(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates LSTM model by defining all layers with activation functions,
@@ -74,6 +106,7 @@ def lstm(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -82,42 +115,19 @@ def lstm(
     model = keras.Sequential()
     model.add(
         LSTM(
-            40,
-            activation='relu',
+            layer_config['layer0'][0],
+            activation=layer_config['layer0'][1],
             return_sequences=True,
             input_shape=input_shape))
-    model.add(LSTM(50, activation='relu', return_sequences=True))
-    model.add(LSTM(50, activation='relu'))
-    model.add(Dense(output_shape))
-    model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
-
-    return model
-
-
-def gru(
-        optimizer: str,
-        loss: str,
-        metrics: str,
-        input_shape: tuple,
-        output_shape: int) -> object:
-    '''Creates GRU model by defining all layers with activation functions,
-        optimizer, loss function and evaluation metrics.
-        Parameters:
-            optimizer (str): Optimization algorithm.
-            loss (str): Loss function.
-            metrics (str): Performance measurement.
-            input_shape (tuple): Time series input shape.
-            ouput_shape (int): Time series output shape.
-        Returns:
-            model (object): Returns compiled Keras model.
-    '''
-    model = keras.Sequential()
-    model.add(GRU(40,
-                  activation='relu',
-                  return_sequences=True,
-                  input_shape=input_shape))
-    model.add(GRU(50, activation='relu', return_sequences=True))
-    model.add(GRU(50, activation='relu'))
+    model.add(
+        LSTM(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1],
+            return_sequences=True))
+    model.add(
+        LSTM(
+            layer_config['layer2'][0],
+            activation=layer_config['layer2'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -128,6 +138,7 @@ def cnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates the CNN model by defining all layers with activation functions,
@@ -136,6 +147,7 @@ def cnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -144,14 +156,56 @@ def cnn(
     model = keras.Sequential()
     model.add(
         Conv1D(
-            filters=64,
-            kernel_size=1,
-            activation='relu',
+            filters=layer_config['layer0'][0],
+            kernel_size=layer_config['layer0'][1],
+            activation=layer_config['layer0'][2],
             input_shape=input_shape))
-    model.add(Conv1D(filters=32, kernel_size=1, activation='relu'))
-    model.add(MaxPooling1D(pool_size=2))
+    model.add(
+        Conv1D(
+            filters=layer_config['layer1'][0],
+            kernel_size=layer_config['layer1'][1],
+            activation=layer_config['layer1'][2]))
+    model.add(MaxPooling1D(pool_size=layer_config['layer2']))
     model.add(Flatten())
-    model.add(Dense(50, activation='relu'))
+    model.add(
+        Dense(
+            layer_config['layer3'][0],
+            activation=layer_config['layer3'][1]))
+    model.add(Dense(output_shape))
+    model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
+
+    return model
+
+
+def gru(
+        optimizer: str,
+        loss: str,
+        metrics: str,
+        layer_config: dict,
+        input_shape: tuple,
+        output_shape: int) -> object:
+    '''Creates GRU model by defining all layers with activation functions,
+        optimizer, loss function and evaluation metrics.
+        Parameters:
+            optimizer (str): Optimization algorithm.
+            loss (str): Loss function.
+            metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
+            input_shape (tuple): Time series input shape.
+            ouput_shape (int): Time series output shape.
+        Returns:
+            model (object): Returns compiled Keras model.
+    '''
+    model = keras.Sequential()
+    model.add(GRU(layer_config['layer0'][0],
+                  activation=layer_config['layer0'][1],
+                  return_sequences=True,
+                  input_shape=input_shape))
+    model.add(GRU(layer_config['layer1'][0],
+                  activation=layer_config['layer1'][1],
+                  return_sequences=True))
+    model.add(GRU(layer_config['layer2'][0],
+                  activation=layer_config['layer2'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -162,6 +216,7 @@ def birnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates a bidirectional RNN model by defining all layers with activation
@@ -170,6 +225,7 @@ def birnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -179,11 +235,14 @@ def birnn(
     model.add(
         Bidirectional(
             SimpleRNN(
-                50,
-                activation='relu',
+                layer_config['layer0'][0],
+                activation=layer_config['layer0'][1],
                 return_sequences=True),
             input_shape=input_shape))
-    model.add(SimpleRNN(50, activation='relu'))
+    model.add(
+        SimpleRNN(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -194,6 +253,7 @@ def bilstm(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates a bidirectional LSTM model by defining all layers with activation
@@ -202,6 +262,7 @@ def bilstm(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -211,11 +272,14 @@ def bilstm(
     model.add(
         Bidirectional(
             LSTM(
-                50,
-                activation='relu',
+                layer_config['layer0'][0],
+                activation=layer_config['layer0'][1],
                 return_sequences=True),
             input_shape=input_shape))
-    model.add(LSTM(50, activation='relu'))
+    model.add(
+        LSTM(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -226,6 +290,7 @@ def bigru(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates a bidirectional GRU model by defining all layers with activation
@@ -234,15 +299,22 @@ def bigru(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
             model (object): Returns compiled Keras model.
     '''
     model = keras.Sequential()
-    model.add(Bidirectional(GRU(50, activation='relu',
-              return_sequences=True), input_shape=input_shape))
-    model.add(GRU(50, activation='relu'))
+    model.add(
+        Bidirectional(
+            GRU(
+                layer_config['layer0'][0],
+                activation=layer_config['layer0'][1],
+                return_sequences=True),
+            input_shape=input_shape))
+    model.add(GRU(layer_config['layer1'][0],
+                  activation=layer_config['layer1'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -253,6 +325,7 @@ def encdec_rnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int,
         repeat: int) -> object:
@@ -262,6 +335,7 @@ def encdec_rnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -270,14 +344,25 @@ def encdec_rnn(
     model = keras.Sequential()
     model.add(
         SimpleRNN(
-            100,
-            activation='relu',
+            layer_config['layer0'][0],
+            activation=layer_config['layer0'][1],
             return_sequences=True,
             input_shape=input_shape))
-    model.add(SimpleRNN(50, activation='relu'))
+    model.add(
+        SimpleRNN(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1]))
     model.add(RepeatVector(repeat))
-    model.add(SimpleRNN(50, activation='relu', return_sequences=True))
-    model.add(SimpleRNN(100, activation='relu', return_sequences=True))
+    model.add(
+        SimpleRNN(
+            layer_config['layer2'][0],
+            activation=layer_config['layer2'][1],
+            return_sequences=True))
+    model.add(
+        SimpleRNN(
+            layer_config['layer3'][0],
+            activation=layer_config['layer3'][1],
+            return_sequences=True))
     model.add(TimeDistributed(Dense(output_shape)))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -288,6 +373,7 @@ def encdec_lstm(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int,
         repeat: int) -> object:
@@ -297,6 +383,7 @@ def encdec_lstm(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -305,14 +392,25 @@ def encdec_lstm(
     model = keras.Sequential()
     model.add(
         LSTM(
-            100,
-            activation='relu',
+            layer_config['layer0'][0],
+            activation=layer_config['layer0'][1],
             return_sequences=True,
             input_shape=input_shape))
-    model.add(LSTM(50, activation='relu'))
+    model.add(
+        LSTM(
+            layer_config['layer1'][0],
+            activation=layer_config['layer1'][1]))
     model.add(RepeatVector(repeat))
-    model.add(LSTM(50, activation='relu', return_sequences=True))
-    model.add(LSTM(100, activation='relu', return_sequences=True))
+    model.add(
+        LSTM(
+            layer_config['layer2'][0],
+            activation=layer_config['layer2'][1],
+            return_sequences=True))
+    model.add(
+        LSTM(
+            layer_config['layer3'][0],
+            activation=layer_config['layer3'][1],
+            return_sequences=True))
     model.add(TimeDistributed(Dense(output_shape)))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -323,6 +421,7 @@ def encdec_cnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int,
         repeat: int) -> object:
@@ -333,6 +432,7 @@ def encdec_cnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -341,16 +441,24 @@ def encdec_cnn(
     model = keras.Sequential()
     model.add(
         Conv1D(
-            filters=64,
-            kernel_size=1,
-            activation='relu',
+            layer_config['layer0'][0],
+            kernel_size=layer_config['layer0'][1],
+            activation=layer_config['layer0'][2],
             input_shape=input_shape))
-    model.add(Conv1D(filters=32, kernel_size=1, activation='relu'))
-    model.add(MaxPooling1D(pool_size=2))
+    model.add(
+        Conv1D(
+            filters=layer_config['layer1'][0],
+            kernel_size=layer_config['layer1'][1],
+            activation=layer_config['layer0'][2]))
+    model.add(MaxPooling1D(pool_size=layer_config['layer2']))
     model.add(Flatten())
     model.add(RepeatVector(repeat))
-    model.add(GRU(50, activation='relu', return_sequences=True))
-    model.add(GRU(100, activation='relu', return_sequences=True))
+    model.add(GRU(layer_config['layer3'][0],
+                  activation=layer_config['layer3'][1],
+                  return_sequences=True))
+    model.add(GRU(layer_config['layer4'][0],
+                  activation=layer_config['layer4'][1],
+                  return_sequences=True))
     model.add(TimeDistributed(Dense(output_shape)))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -361,6 +469,7 @@ def encdec_gru(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int,
         repeat: int) -> object:
@@ -370,20 +479,26 @@ def encdec_gru(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
             model (object): Returns compiled Keras model.
     '''
     model = keras.Sequential()
-    model.add(GRU(100,
-                  activation='relu',
+    model.add(GRU(layer_config['layer0'][0],
+                  activation=layer_config['layer0'][1],
                   return_sequences=True,
                   input_shape=input_shape))
-    model.add(GRU(50, activation='relu'))
+    model.add(GRU(layer_config['layer1'][0],
+                  activation=layer_config['layer1'][1]))
     model.add(RepeatVector(repeat))
-    model.add(GRU(50, activation='relu', return_sequences=True))
-    model.add(GRU(100, activation='relu', return_sequences=True))
+    model.add(GRU(layer_config['layer2'][0],
+                  activation=layer_config['layer2'][1],
+                  return_sequences=True))
+    model.add(GRU(layer_config['layer3'][0],
+                  activation=layer_config['layer3'][1],
+                  return_sequences=True))
     model.add(TimeDistributed(Dense(output_shape)))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -394,6 +509,7 @@ def cnnrnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-RNN hybrid model by defining all layers with activation
@@ -402,6 +518,7 @@ def cnnrnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -411,20 +528,27 @@ def cnnrnn(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
-    model.add(SimpleRNN(50, activation='relu', return_sequences=True))
-    model.add(SimpleRNN(25, activation='relu'))
+    model.add(
+        SimpleRNN(
+            layer_config['layer3'][0],
+            activation=layer_config['layer3'][1],
+            return_sequences=True))
+    model.add(
+        SimpleRNN(
+            layer_config['layer4'][0],
+            activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -435,6 +559,7 @@ def cnnlstm(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-LSTM hybrid model by defining all layers with activation
@@ -443,6 +568,7 @@ def cnnlstm(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -452,20 +578,27 @@ def cnnlstm(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
-    model.add(LSTM(50, activation='relu', return_sequences=True))
-    model.add(LSTM(25, activation='relu'))
+    model.add(
+        LSTM(
+            layer_config['layer3'][0],
+            activation=layer_config['layer3'][1],
+            return_sequences=True))
+    model.add(
+        LSTM(
+            layer_config['layer4'][0],
+            activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -476,6 +609,7 @@ def cnngru(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-GRU hybrid model by defining all layers with activation
@@ -484,6 +618,7 @@ def cnngru(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -493,20 +628,23 @@ def cnngru(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
-    model.add(GRU(50, activation='relu', return_sequences=True))
-    model.add(GRU(25, activation='relu'))
+    model.add(GRU(layer_config['layer3'][0],
+                  activation=layer_config['layer3'][1],
+                  return_sequences=True))
+    model.add(GRU(layer_config['layer4'][0],
+                  activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -517,6 +655,7 @@ def cnnbirnn(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-BI-RNN hybrid model by defining all layers with activation
@@ -525,6 +664,7 @@ def cnnbirnn(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -534,25 +674,28 @@ def cnnbirnn(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
     model.add(
         Bidirectional(
             SimpleRNN(
-                50,
-                activation='relu',
+                layer_config['layer3'][0],
+                activation=layer_config['layer3'][1],
                 return_sequences=True)))
-    model.add(SimpleRNN(25, activation='relu'))
+    model.add(
+        SimpleRNN(
+            layer_config['layer4'][0],
+            activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -563,6 +706,7 @@ def cnnbilstm(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-BI-LSTM hybrid model by defining all layers with activation
@@ -571,6 +715,7 @@ def cnnbilstm(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -580,25 +725,28 @@ def cnnbilstm(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
     model.add(
         Bidirectional(
             LSTM(
-                50,
-                activation='relu',
+                layer_config['layer3'][0],
+                activation=layer_config['layer3'][1],
                 return_sequences=True)))
-    model.add(LSTM(25, activation='relu'))
+    model.add(
+        LSTM(
+            layer_config['layer4'][0],
+            activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
 
@@ -609,6 +757,7 @@ def cnnbigru(
         optimizer: str,
         loss: str,
         metrics: str,
+        layer_config: dict,
         input_shape: tuple,
         output_shape: int) -> object:
     '''Creates CNN-BI-GRU hybrid model by defining all layers with activation
@@ -617,6 +766,7 @@ def cnnbigru(
             optimizer (str): Optimization algorithm.
             loss (str): Loss function.
             metrics (str): Performance measurement.
+            layer_config (dict): Adjust neurons and activation functions.
             input_shape (tuple): Time series input shape.
             ouput_shape (int): Time series output shape.
         Returns:
@@ -626,20 +776,26 @@ def cnnbigru(
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=64,
-                kernel_size=1,
-                activation='relu'),
+                filters=layer_config['layer0'][0],
+                kernel_size=layer_config['layer0'][1],
+                activation=layer_config['layer0'][2]),
             input_shape=input_shape))
     model.add(
         TimeDistributed(
             Conv1D(
-                filters=32,
-                kernel_size=1,
-                activation='relu')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
+                filters=layer_config['layer1'][0],
+                kernel_size=layer_config['layer1'][1],
+                activation=layer_config['layer1'][2])))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=layer_config['layer2'])))
     model.add(TimeDistributed(Flatten()))
-    model.add(Bidirectional(GRU(50, activation='relu', return_sequences=True)))
-    model.add(GRU(25, activation='relu'))
+    model.add(
+        Bidirectional(
+            GRU(
+                layer_config['layer3'][0],
+                activation=layer_config['layer3'][1],
+                return_sequences=True)))
+    model.add(GRU(layer_config['layer4'][0],
+                  activation=layer_config['layer4'][1]))
     model.add(Dense(output_shape))
     model.compile(optimizer=optimizer, loss=loss, metrics=metrics)